Add files via upload
Mouse movement to LUA converter with a GUI. A bit more complicated than my original script but performs the exact same function, but with a more user friendly GUI. 

Allows you to set the desired hotkey, and save location of the LUA file. Also added sequential naming conventions for when you spam the script.